Undo fixes - the problem is with mktime on 32-bit systems. Remove tests that break on 32-bit systems.

+++ src/funcs.c
@@ -3758,9 +3758,9 @@ solstice_equinox_for_year(int y, int which)
     }
 
     j -= 2447892.50000;  /* This is the Julian date of midnight, 1 Jan 1990 UTC */
-    long long dse = (long long) j;
+    int dse = (int) j;
 
-    long long min = (long long) floor((j - (double) dse) * MINUTES_PER_DAY);
+    int min = (int) floor((j - (double) dse) * MINUTES_PER_DAY);
     int ret;
 
     /* Convert from UTC to local time */
